refactor(crud): split membershipsCrud and standingBookingsCrud into packages
Both monolithic CRUD modules (1204 lines and growing) are split into
focused submodules under app/crud/memberships/ and app/crud/standing_bookings/.
The original module names are kept as thin shims that re-export the public
API so call sites do not need to change.

This refactor is the baseline that the payments scaling work in subsequent
commits builds on.
