+++ app/crud/memberships/__init__.py
@@ -1,0 +1,1 @@
+"""Membership CRUD helpers split by domain."""
