+++ app/crud/standing_bookings/__init__.py
@@ -1,0 +1,1 @@
+"""Standing bookings CRUD split by domain."""
